+++ linuxkm/lkcapi_ecdsa_glue.c
@@ -749,15 +749,15 @@ static int linuxkm_test_ecdsa_nist_driver(const char * driver,
          * in kernel crypto/testmgr.c, and the kernel will block
          * its allocation if fips_enabled is set. */
         if ((PTR_ERR(tfm) == -ENOENT) && fips_enabled) {
-            pr_info("info: skipping unsupported akcipher algorithm %s: %ld\n",
-                    driver, PTR_ERR(tfm));
+            pr_info("info: skipping unsupported akcipher algorithm %s: %d\n",
+                    driver, (int)PTR_ERR(tfm));
             test_rc = NOT_COMPILED_IN;
         }
         else
         #endif
         {
-            pr_err("error: allocating akcipher algorithm %s failed: %ld\n",
-                   driver, PTR_ERR(tfm));
+            pr_err("error: allocating akcipher algorithm %s failed: %d\n",
+                   driver, (int)PTR_ERR(tfm));
             if (PTR_ERR(tfm) == -ENOMEM)
                 test_rc = MEMORY_E;
             else
